+++ vmm/src/vmm_config/instance_info.rs
@@ -38,6 +38,8 @@ pub struct InstanceInfo {
     pub id: String,
     /// The state of the microVM.
     pub state: InstanceState,
+    /// The version of the VMM that runs the microVM.
+    pub vmm_version: String,
 }
 
 /// Errors associated with starting the instance.